pyproject.toml
[project]
name = "autosrt"
version = "0.1.0"
description = ""
authors = [
    {name = "Sakuya0229",email = "a1301145@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.0"
dependencies = [
    "srt (>=3.5.3,<4.0.0)",
    "datetime (>=5.5,<6.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "torchaudio (>=2.8.0,<3.0.0)",
    "pyside6 (>=6.9.2,<7.0.0)",
    "pyinstaller (>=6.15.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}
torchvision = {source = "pytorch-gpu"}
torchaudio = {source = "pytorch-gpu"}
